major refactoring of serialization classes / adapters
