+++ bigdataviewer.properties
@@ -1,8 +1,8 @@
 #
-#Fri Feb 05 10:33:13 CET 2021
+#Fri Apr 30 16:10:12 CEST 2021
 scale-bar-bg-color=-16724788
 show-text-overlay=true
-show-multibox-overlay=true
+show-multibox-overlay=false
 show-scale-bar-in-movie=true
 show-scale-bar=true
 scale-bar-color=-1
