+++ plugins/bdvpgsettings/bdvpg.tree.actions.txt
@@ -1,0 +1,37 @@
+[
+  "BDV - Show Sources",
+  "BDV - Show Sources (new Bdv window)",
+  "BDV - Zoom to source",
+  "BDV - Show Sources In Multiple BDV Windows",
+  "BDV - Set Number Of Timepoints",
+  "BDV - Create empty BDV window",
+  "BDV - Create Orthogonal Views",
+  "BDV - Screenshot",
+  "BDV - Synchronize Views",
+  "BDV - Remove Sources From BDV",
+  "PopupLine",
+  "Create Empty BVV Frame",
+  "Show Sources in BVV",
+  "PopupLine",
+  "Set Sources Color",
+  "Set Sources Brightness",
+  "Make Sources Invisible",
+  "Make Sources Visible",
+  "Set Sources Projection Mode",
+  "Create New Source (Set LUT)",
+  "Create New Source (Set Color)",
+  "PopupLine",
+  "Manual Sources Transformation",
+  "Basic Transformation",
+  "Make Global Source Group",
+  "PopupLine",
+  "Save BDVDataset",
+  "Inspect Sources",
+  "PopupLine",
+  "Add Metadata To Sources",
+  "Make Metadata Filter Node",
+  "PopupLine",
+  "Export Sources to XML/HDF5 Spimdataset",
+  "Delete Sources",
+  "PopupLine"
+]
