+++ src/main/java/sc/fiji/bdvpg/scijava/adapter/bdv/DefaultBdvSupplierAdapter.java
@@ -1,7 +1,10 @@
-package sc.fiji.bdvpg.services.serializers.bdv;
+package sc.fiji.bdvpg.scijava.adapter.bdv;
 
 import com.google.gson.*;
 import org.scijava.plugin.Plugin;
+import sc.fiji.bdvpg.bdv.supplier.DefaultBdvSupplier;
+import sc.fiji.bdvpg.bdv.supplier.IBdvSupplier;
+import sc.fiji.bdvpg.bdv.supplier.SerializableBdvOptions;
 import sc.fiji.serializers.IClassRuntimeAdapter;
 
 import java.lang.reflect.Type;
