+++ src/main/java/sc/fiji/bdvpg/scijava/adapter/transform/WrappedIterativeInvertibleRealTransformAdapter.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.bdvpg.services.serializers.plugins;
+package sc.fiji.bdvpg.scijava.adapter.transform;
 
 import com.google.gson.*;
 import net.imglib2.realtransform.RealTransform;
